Added an extension to look for the environment value XCTestSessionIdentifier and return true #if DEBUG and XCTestSessionIdentifier is nil which prevents a problem of running deleteAll() and createSampleData() when we are debugging

+++ UltimatePortfolio/Controllers/DataController.swift
@@ -42,10 +42,10 @@ public final class DataController: ObservableObject {
             if let error = error { fatalError(error.localizedDescription) }
         }
 
-        #if DEBUG
-        self.deleteAll()
-        try? self.createSampleData()
-        #endif
+        if ProcessInfo.processInfo.isDebugNotTesting() {
+            self.deleteAll()
+            try? self.createSampleData()
+        }
     }
 
     /// Saves our Core Data context iff (if and only if) there are changes. This silently ignores
